Create Java Static Analysis signature
Begin to take advantage of the new Java static analysis. I intend to expand this out further in time but this is a start (although exploit kits and things have moved on from Java now but just in case old exploits turn up and to try and detect any new use that might occur). References are included in the signature.